Update airports table

+++ sql/create_tables.sql
@@ -7,8 +7,8 @@ CREATE TABLE IF NOT EXISTS airports (
     name VARCHAR(255) NOT NULL,
     city VARCHAR(255) NOT NULL,
     country VARCHAR(50) NOT NULL,
-    latitude DECIMAL(10, 8) NOT NULL,
-    longitude DECIMAL(11, 8) NOT NULL
+    latitude DOUBLE PRECISION NOT NULL,
+    longitude DOUBLE PRECISION NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS users (
